3208 add candidate_office and candidate_state fields to commitee_account model
