+++ django-backend/fecfiler/contacts/shared_models.py
@@ -1,0 +1,14 @@
+from django.db import models
+
+
+class ContactType(models.TextChoices):
+    COMMITTEE = "COM"
+    ORGANIZATION = "ORG"
+    INDIVIDUAL = "IND"
+    CANDIDATE = "CAN"
+
+
+class CandidateOffice(models.TextChoices):
+    HOUSE = "H"
+    SENATE = "S"
+    PRESIDENTIAL = "P"
