+++ django-backend/fecfiler/web_services/dot_fec/tests/dot_fec_record_testing/test_schedule_b.py
@@ -6,6 +6,7 @@ from fecfiler.committee_accounts.models import CommitteeAccount
 from fecfiler.reports.tests.utils import create_form3x
 from fecfiler.transactions.tests.utils import create_schedule_b
 from fecfiler.contacts.models import Contact
+from fecfiler.contacts.shared_models import ContactType
 from datetime import datetime
 from fecfiler.web_services.models import UploadSubmission
 import structlog
@@ -18,7 +19,7 @@ class DotFECScheduleBTestCase(TestCase):
         self.committee = CommitteeAccount.objects.create(committee_id="C00000000")
 
         self.organization = Contact.objects.create(
-            type=Contact.ContactType.ORGANIZATION,
+            type=ContactType.ORGANIZATION,
             name="Test Org",
             committee_id=self.committee.id,
             street_1="5678 Road St",
@@ -29,7 +30,7 @@ class DotFECScheduleBTestCase(TestCase):
         )
 
         self.candidate = Contact.objects.create(
-            type=Contact.ContactType.CANDIDATE,
+            type=ContactType.CANDIDATE,
             last_name="Candidate last",
             first_name="Candidate first",
             middle_name="Candidate middle",
@@ -43,7 +44,7 @@ class DotFECScheduleBTestCase(TestCase):
         )
 
         self.individual = Contact.objects.create(
-            type=Contact.ContactType.INDIVIDUAL,
+            type=ContactType.INDIVIDUAL,
             last_name="last name",
             first_name="First name",
             committee_id=self.committee.committee_id,
