+++ django-backend/fecfiler/web_services/dot_fec/tests/dot_fec_record_testing/test_schedule_c.py
@@ -7,6 +7,7 @@ from fecfiler.committee_accounts.models import CommitteeAccount
 from fecfiler.reports.tests.utils import create_form3x
 from fecfiler.transactions.tests.utils import create_loan
 from fecfiler.contacts.models import Contact
+from fecfiler.contacts.shared_models import ContactType
 from datetime import datetime
 from fecfiler.web_services.models import UploadSubmission
 import structlog
@@ -31,7 +32,7 @@ class DotFECScheduleCTestCase(TestCase):
         self.f3x.save()
 
         self.individual = Contact.objects.create(
-            type=Contact.ContactType.INDIVIDUAL,
+            type=ContactType.INDIVIDUAL,
             last_name="last name",
             first_name="First name",
             committee_id=self.committee.committee_id,
@@ -46,7 +47,7 @@ class DotFECScheduleCTestCase(TestCase):
         )
 
         self.candidate = Contact.objects.create(
-            type=Contact.ContactType.CANDIDATE,
+            type=ContactType.CANDIDATE,
             last_name="Candidate last",
             first_name="Candidate first",
             middle_name="Candidate middle",
@@ -60,7 +61,7 @@ class DotFECScheduleCTestCase(TestCase):
         )
 
         self.organization = Contact.objects.create(
-            type=Contact.ContactType.ORGANIZATION,
+            type=ContactType.ORGANIZATION,
             name="Test Org",
             committee_id=self.committee.id,
             street_1="5678 Road St",
