+++ django-backend/fecfiler/web_services/dot_fec/tests/dot_fec_record_testing/test_schedule_c1.py
@@ -7,6 +7,7 @@ from fecfiler.committee_accounts.models import CommitteeAccount
 from fecfiler.reports.tests.utils import create_form3x
 from fecfiler.transactions.tests.utils import create_loan_from_bank
 from fecfiler.contacts.models import Contact
+from fecfiler.contacts.shared_models import ContactType
 from datetime import datetime
 from fecfiler.web_services.models import UploadSubmission
 import structlog
@@ -31,7 +32,7 @@ class DotFECScheduleC1TestCase(TestCase):
         self.f3x.save()
 
         self.organization = Contact.objects.create(
-            type=Contact.ContactType.ORGANIZATION,
+            type=ContactType.ORGANIZATION,
             name="Test Org",
             committee_id=self.committee.id,
             street_1="5678 Road St",
