+++ django-backend/fecfiler/web_services/dot_fec/tests/dot_fec_record_testing/test_schedule_d.py
@@ -7,6 +7,7 @@ from fecfiler.committee_accounts.models import CommitteeAccount
 from fecfiler.reports.tests.utils import create_form3x
 from fecfiler.transactions.tests.utils import create_debt
 from fecfiler.contacts.models import Contact
+from fecfiler.contacts.shared_models import ContactType
 from datetime import datetime
 from fecfiler.web_services.models import UploadSubmission
 from fecfiler.transactions.aggregation import process_aggregation_for_debts
@@ -32,7 +33,7 @@ class DotFECScheduleDTestCase(TestCase):
         self.f3x.save()
 
         self.individual = Contact.objects.create(
-            type=Contact.ContactType.INDIVIDUAL,
+            type=ContactType.INDIVIDUAL,
             last_name="last name",
             first_name="First name",
             committee_account_id=self.committee.id,
@@ -47,7 +48,7 @@ class DotFECScheduleDTestCase(TestCase):
         )
 
         self.organization = Contact.objects.create(
-            type=Contact.ContactType.ORGANIZATION,
+            type=ContactType.ORGANIZATION,
             name="Test Org",
             committee_account_id=self.committee.id,
             street_1="5678 Road St",
